requirements.txt
# Core dependencies
datasets>=2.14.0
transformers>=4.44.0
accelerate>=0.34.0
evaluate>=0.4.2
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
torch>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code quality
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0

# Data version control
dvc>=3.0.0
dvc-s3>=3.0.0


# Experiment tracking
mlflow>=2.10.0
